Moved Enhanced Editor methods to shared file
